[react-interactions] Refactor TabFocusController (#16768)
